Move registration of basic build file constructs.
They clearly are not part of any backend, and they aren't
tasks (so they don't belong in core_tasks), so they
 now go in build_graph, which is where almost all the code
that implements them lives.

Moved wrapped_globs.py from backend/core to source, as that
seemed like a more apt place for it than build_graph.

They deal with finding source files on disk, so the source
package is a perfect place for them.
