+++ contrib/go/src/python/pants/contrib/go/targets/BUILD
@@ -29,9 +29,9 @@ python_library(
   sources=['go_local_source.py'],
   dependencies=[
     ':go_target',
-    'src/python/pants/backend/core:wrapped_globs',
     'src/python/pants/base:build_environment',
     'src/python/pants/base:payload',
+    'src/python/pants/source',
   ],
 )
