+++ src/python/pants/backend/core/BUILD
@@ -7,9 +7,6 @@ python_library(
   sources = ['register.py'],
   dependencies = [
     'src/python/pants/backend/core/tasks:all',
-    'src/python/pants/base:build_environment',
-    'src/python/pants/base:source_root',
-    'src/python/pants/build_graph',
     'src/python/pants/goal:task_registrar',
     ':core',
   ]
@@ -21,11 +18,8 @@ python_library(
     'wrapped_globs.py',
   ],
   dependencies = [
-    '3rdparty/python/twitter/commons:twitter.common.dirutil',
-    '3rdparty/python:six',
-    'src/python/pants/base:build_environment',
-    'src/python/pants/base:build_manual',
-    'src/python/pants/util:memo',
+    'src/python/pants/base:deprecated',
+    'src/python/pants/source',
   ],
 )
 
@@ -35,11 +29,7 @@ python_library(
     'from_target.py',
   ],
   dependencies = [
-    '3rdparty/python:six',
-    'src/python/pants/backend/core/tasks:all',
-    'src/python/pants/base:build_environment',
+    'src/python/pants/base:deprecated',
     'src/python/pants/build_graph',
-    'src/python/pants/util:dirutil',
-    ':wrapped_globs',
   ]
 )
