+++ src/python/pants/base/payload_field.py
@@ -12,9 +12,9 @@ from hashlib import sha1
 
 from twitter.common.collections import OrderedSet
 
-from pants.backend.core import wrapped_globs
 from pants.base.build_environment import get_buildroot
 from pants.base.validation import assert_list
+from pants.source import wrapped_globs
 from pants.util.meta import AbstractClass
