+++ src/python/pants/bin/BUILD
@@ -19,6 +19,7 @@ python_library(
     'src/python/pants/base:scm_build_file',
     'src/python/pants/base:workunit',
     'src/python/pants/build_graph',
+    'src/python/pants/core_tasks',
     'src/python/pants/engine',
     'src/python/pants/goal',
     'src/python/pants/goal:context',
@@ -47,7 +48,6 @@ target(
     'src/python/pants/backend/jvm:plugin',
     'src/python/pants/backend/project_info:plugin',
     'src/python/pants/backend/python:plugin',
-    'src/python/pants/core_tasks',
   ],
 )
