+++ src/python/pants/build_graph/target.py
@@ -11,9 +11,7 @@ from hashlib import sha1
 
 from six import string_types
 
-from pants.backend.core.wrapped_globs import FilesetWithSpec
 from pants.base.build_environment import get_buildroot
-from pants.base.deprecated import deprecated
 from pants.base.exceptions import TargetDefinitionException
 from pants.base.fingerprint_strategy import DefaultFingerprintStrategy
 from pants.base.hash_utils import hash_all
@@ -24,6 +22,7 @@ from pants.build_graph.address import Address, Addresses
 from pants.build_graph.target_addressable import TargetAddressable
 from pants.option.custom_types import dict_option
 from pants.source.source_root import SourceRootConfig
+from pants.source.wrapped_globs import FilesetWithSpec
 from pants.subsystem.subsystem import Subsystem
 from pants.util.memo import memoized_property
