+++ src/python/pants/engine/exp/BUILD
@@ -93,5 +93,6 @@ python_library(
   dependencies=[
     ':addressable',
     ':configuration',
+    'src/python/pants/source',
   ]
 )
