+++ src/python/pants/source/BUILD
@@ -6,7 +6,9 @@ python_library(
   sources=globs('*.py'),
   dependencies=[
     '3rdparty/python:six',
+    '3rdparty/python/twitter/commons:twitter.common.dirutil',
     'src/python/pants/base:build_environment',
+    'src/python/pants/base:build_manual',
     'src/python/pants/option',
     'src/python/pants/subsystem',
     'src/python/pants/util:memo',
