+++ tests/python/pants_test/backend/codegen/tasks/BUILD
@@ -43,7 +43,7 @@ python_tests(
     'src/python/pants/backend/codegen:plugin',
     'src/python/pants/backend/codegen/targets:java',
     'src/python/pants/backend/codegen/tasks:jaxb_gen',
-    'src/python/pants/backend/core:plugin',
+    'src/python/pants/build_graph',
     'src/python/pants/util:contextutil',
     'tests/python/pants_test/tasks:task_test_base',
   ],
@@ -55,7 +55,7 @@ python_tests(
   dependencies = [
     'src/python/pants/backend/codegen/tasks:protobuf_gen',
     'src/python/pants/backend/codegen:plugin',
-    'src/python/pants/backend/core:plugin',
+    'src/python/pants/build_graph',
     'src/python/pants/util:contextutil',
     'tests/python/pants_test/tasks:task_test_base',
   ],
@@ -88,7 +88,7 @@ python_tests(
     'src/python/pants/backend/codegen/targets:java',
     'src/python/pants/backend/codegen/tasks:simple_codegen_task',
     'src/python/pants/backend/codegen:plugin',
-    'src/python/pants/backend/core:plugin',
+    'src/python/pants/build_graph',
     'src/python/pants/util:contextutil',
     'src/python/pants/util:dirutil',
     'tests/python/pants_test/tasks:task_test_base',
@@ -103,7 +103,6 @@ python_tests(
     'src/python/pants/backend/codegen/targets:java',
     'src/python/pants/backend/codegen/tasks:wire_gen',
     'src/python/pants/backend/codegen:plugin',
-    'src/python/pants/backend/core:plugin',
     'src/python/pants/backend/jvm/targets:jvm',
     'src/python/pants/base:exceptions',
     'src/python/pants/base:revision',
