+++ tests/python/pants_test/backend/codegen/tasks/test_jaxb_gen.py
@@ -10,7 +10,7 @@ import os
 from pants.backend.codegen.register import build_file_aliases as register_codegen
 from pants.backend.codegen.targets.jaxb_library import JaxbLibrary
 from pants.backend.codegen.tasks.jaxb_gen import JaxbGen
-from pants.backend.core.register import build_file_aliases as register_core
+from pants.build_graph.register import build_file_aliases as register_core
 from pants_test.tasks.task_test_base import TaskTestBase
