--- tests/python/pants_test/backend/core/BUILD
@@ -1,11 +1,0 @@
-# Copyright 2014 Pants project contributors (see CONTRIBUTORS.md).
-# Licensed under the Apache License, Version 2.0 (see LICENSE).
-
-python_tests(
-  name = 'wrapped_globs',
-  sources = ['test_wrapped_globs.py'],
-  dependencies = [
-    'src/python/pants/backend/core',
-    'tests/python/pants_test:base_test',
-  ]
-)
