+++ tests/python/pants_test/backend/jvm/targets/BUILD
@@ -23,10 +23,10 @@ python_tests(
   name='jvm_app',
   sources=['test_jvm_app.py'],
   dependencies=[
-    'src/python/pants/backend/core:wrapped_globs',
     'src/python/pants/backend/jvm/targets:jvm',
     'src/python/pants/base:exceptions',
     'src/python/pants/build_graph',
+    'src/python/pants/source',
     'tests/python/pants_test:base_test',
   ]
 )
