+++ tests/python/pants_test/source/BUILD
@@ -10,3 +10,14 @@ python_tests(
     'tests/python/pants_test/subsystem:subsystem_utils'
   ]
 )
+
+python_tests(
+  name = 'wrapped_globs',
+  sources = ['test_wrapped_globs.py'],
+  dependencies = [
+    'src/python/pants/backend/jvm/targets:java',
+    'src/python/pants/build_graph',
+    'src/python/pants/source',
+    'tests/python/pants_test:base_test',
+  ]
+)
